* upstream/master:
  MRG, ENH: Add method to project onto max power ori (#7883)
  WIP: Warn if untested NIRX device (#7905)
  MRG, BUG: Fix bug with volume morph and subject_to!="fsaverage" (#7896)
  MRG, MAINT: Clean up use of bool, float, int (#7917)
  ENH: Better error message for incompatible Evoked objects (#7910)
  try to fix nullcontext (#7908)

+++ mne/simulation/evoked.py
@@ -73,7 +73,7 @@ def simulate_evoked(fwd, stc, info, cov, nave=30, iir_filter=None,
     if nave < np.inf:
         noise = _simulate_noise_evoked(evoked, cov, iir_filter, random_state)
         evoked.data += noise.data / math.sqrt(nave)
-        evoked.nave = np.int(nave)
+        evoked.nave = np.int64(nave)
     if cov is not None and cov.get('projs', None):
         evoked.add_proj(cov['projs']).apply_proj()
     return evoked
